chore: Update app.py to bind to all network interfaces
This commit updates the `app.py` file to bind the Flask application to all network interfaces by specifying the `host` parameter as `"0.0.0.0"` and the `port` parameter as `5000`. This change allows the application to be accessible from any IP address.

Note: This commit message follows the established convention of starting with a verb in the imperative mood, followed by a brief description of the changes made.